==> 6) gap glm.Rmd <==
---
title: "6) gap glm"
author: "Emmerson Wilson"
date: "2024-06-17"
output: html_document
---


# load packages
```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  cv,
  lme4,
  AICcmodavg,
  rcompanion,
  brglm2)
  
```

# set working directory
```{r}
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/c_gap_predictions")
```

# data
```{r}
c_env_sites_scale_GM_gap <- read_xlsx("created data/c_env_sites_scale_GM_gap.xlsx")
c_env_sites_scale_TN_gap <- read_xlsx("created data/c_env_sites_scale_TN_gap.xlsx")
```

# glm functions
```{r}
source("4) glm functions.R")
```

# GM

## null
```{r}
#run the model
GM_lm_null_gap <- firths_glm(c_env_sites_scale_GM_gap, "gap", "1")
```

```{r}
# get the summary of the model into a dataframe
GM_lm_null_gap_summary <- as.data.frame(summary(GM_lm_null_gap)$coefficients) %>% 
  mutate(model = "GM_lm_null") 
GM_lm_null_gap_summary <- GM_lm_null_gap_summary %>% 
  add_column("predictor" = rownames(GM_lm_null_gap_summary))

# reorder columns so model name first
GM_lm_null_gap_summary <- GM_lm_null_gap_summary[,c(5,6,1:4)]
GM_lm_null_gap_summary
```

## full
```{r}
# make list of all variables to be included in full model
full_var <- list("FHT", "EVIamp", "EVImed", "ELE", "SLO", "ASP", "LGS", "Moose_new")

# run the model
GM_lm_full_gap <- firths_glm(c_env_sites_scale_GM_gap, "gap", full_var)
summary(GM_lm_full_gap)
```

```{r}
# add summary of model to dataframe
GM_lm_full_gap_summary <- as.data.frame(summary(GM_lm_full_gap)$coefficients) %>% 
  mutate(model = "GM_lm_full") 
GM_lm_full_gap_summary <- GM_lm_full_gap_summary %>% 
  add_column("predictor" = rownames(GM_lm_full_gap_summary))

GM_lm_full_gap_summary <- GM_lm_full_gap_summary[,c(5,6,1:4)]
GM_lm_full_gap_summary
```

## univariate
```{r}
GM_lm_uni_gap <- lapply(full_var, firths_glm, df = c_env_sites_scale_GM_gap, c = "gap")

names(GM_lm_uni_gap) <- full_var # make sure names of objects in list consistent
```

```{r}
# get summary of each univariate model
GM_lm_uni_gap_summary <- lapply(GM_lm_uni_gap, summary)

# add to dataframe
GM_uni_gap_summary_all <- Reduce(full_join,lapply(mapply(cbind, lapply(lapply(GM_lm_uni_gap_summary, coef), as.data.frame), "model"=names(GM_lm_uni_gap_summary), SIMPLIFY=F), rownames_to_column, var = "predictor")) 
GM_uni_gap_summary_all <- GM_uni_gap_summary_all[,c(5,6,1:4)]
GM_uni_gap_summary_all
```

## reduced
```{r}
# find possible reduced model
GM_lm_full_gap2 <- glm(formula = gap ~ FHT + EVIamp + EVImed + ELE + SLO + ASP + LGS + Moose_new, data = c_env_sites_scale_GM_gap, family = binomial(link = "logit"), method = "brglmFit") 

GM_lm_red_gap <- step(GM_lm_full_gap2, direction = "backward", trace = F)
```

```{r}
# add reduced models together to compare
GM_gap_red_models_r <- list()
GM_gap_red_models_r[[names(GM_lm_uni_gap)[1]]] <- GM_lm_uni_gap[[1]]
GM_gap_red_models_r[[names(GM_lm_uni_gap)[2]]] <- GM_lm_uni_gap[[2]]
GM_gap_red_models_r[["GM_lm_red_gap"]] <- GM_lm_red_gap
Modnames_GM_gap_red <- paste(names(GM_gap_red_models_r))

# get AICc, R2, Log likelihood
# requires own functions to work with brglm2

# for R2
NagelkerkeR2_own <- function(model, null, N) {
1 - ((((exp(logLik(null)[[1]]))/exp(logLik(model)[[1]])))^(2/N))/(1-exp(logLik(null)[[1]])^(2/N))
}

GM_N <- length(c_env_sites_scale_GM_gap$unique_id)

R2_GM_gap_red <- data.frame(lapply(GM_gap_red_models_r, NagelkerkeR2_own, null = GM_lm_null_gap, N= GM_N)) %>% 
  pivot_longer(cols = c(1:3), values_to = "R2", names_to = "Modnames")

## Log likelihood
LL_GM_gap_red <- data.frame(lapply(lapply(GM_gap_red_models_r, logLik), function(x) x[[1]])) %>% 
  pivot_longer(cols = c(1:3), values_to = "LL", names_to = "Modnames")

## AICc
AICc_GM_gap_red <- data.frame(lapply(GM_gap_red_models_r, AICc)) %>% 
  pivot_longer(cols = c(1:3), values_to = "AICc", names_to = "Modnames") 

## put together
aic_R2_LL_GM_gap <- full_join(LL_GM_gap_red, AICc_GM_gap_red) %>% 
  full_join(R2_GM_gap_red) 
```

## compare
```{r}
GM_models_gap_r <- list()

GM_models_gap_r[["GM_lm_null_gap"]] <- GM_lm_null_gap
GM_models_gap_r[["GM_lm_full_gap"]] <- GM_lm_full_gap
GM_models_gap_r[[names(GM_lm_uni_gap)[1]]] <- GM_lm_uni_gap[[1]]
GM_models_gap_r[[names(GM_lm_uni_gap)[2]]] <- GM_lm_uni_gap[[2]]
GM_models_gap_r[[names(GM_lm_uni_gap)[3]]] <- GM_lm_uni_gap[[3]]
GM_models_gap_r[[names(GM_lm_uni_gap)[4]]] <- GM_lm_uni_gap[[4]]
GM_models_gap_r[[names(GM_lm_uni_gap)[5]]] <- GM_lm_uni_gap[[5]]
GM_models_gap_r[[names(GM_lm_uni_gap)[6]]] <- GM_lm_uni_gap[[6]]
GM_models_gap_r[[names(GM_lm_uni_gap)[7]]] <- GM_lm_uni_gap[[7]]
GM_models_gap_r[[names(GM_lm_uni_gap)[8]]] <- GM_lm_uni_gap[[8]]
GM_models_gap_r[["GM_lm_red_gap"]] <- GM_lm_red_gap

Modnames_GM_gap_r <- paste(names(GM_models_gap_r)) # make sure names are same as model name

LL_GM_gap <- data.frame(lapply(lapply(GM_models_gap_r, logLik), function(x) x[[1]])) %>% 
  pivot_longer(cols = c(1:11), values_to = "LL", names_to = "Modnames")

R2_GM_gap <- data.frame(lapply(GM_models_gap_r, NagelkerkeR2_own, null = GM_lm_null_gap, N= GM_N)) %>% 
  pivot_longer(cols = c(1:11), values_to = "R2", names_to = "Modnames")

AICc_GM_gap <- data.frame(lapply(GM_models_gap_r, AICc)) %>% 
  pivot_longer(cols = c(1:11), values_to = "AICc", names_to = "Modnames") 

aic_R2_LL_GM_gap <- full_join(LL_GM_gap, AICc_GM_gap) %>% 
  full_join(R2_GM_gap)
```

# TN

## null
```{r}
#run the model
TN_lm_null_gap <- firths_glm(c_env_sites_scale_TN_gap, "gap", "1")
```

```{r}
# get the summary of the model into a dataframe
TN_lm_null_gap_summary <- as.data.frame(summary(TN_lm_null_gap)$coefficients) %>% 
  mutate(model = "TN_lm_null") 
TN_lm_null_gap_summary <- TN_lm_null_gap_summary %>% 
  add_column("predictor" = rownames(TN_lm_null_gap_summary))

# reorder columns so model name first
TN_lm_null_gap_summary <- TN_lm_null_gap_summary[,c(5,6,1:4)]
TN_lm_null_gap_summary
```

## full
```{r}
# make list of all variables to be included in full model
full_var <- list("FHT", "EVIamp", "EVImed", "ELE", "SLO", "ASP", "LGS", "Moose_new")

# run the model
TN_lm_full_gap <- firths_glm(c_env_sites_scale_TN_gap, "gap", full_var)
summary(TN_lm_full_gap)
```

```{r}
# add summary of model to dataframe
TN_lm_full_gap_summary <- as.data.frame(summary(TN_lm_full_gap)$coefficients) %>% 
  mutate(model = "TN_lm_full") 
TN_lm_full_gap_summary <- TN_lm_full_gap_summary %>% 
  add_column("predictor" = rownames(TN_lm_full_gap_summary))

TN_lm_full_gap_summary <- TN_lm_full_gap_summary[,c(5,6,1:4)]
TN_lm_full_gap_summary
```

## univariate
```{r}
TN_lm_uni_gap <- lapply(full_var, firths_glm, df = c_env_sites_scale_TN_gap, c = "gap")

names(TN_lm_uni_gap) <- full_var # make sure names of objects in list consistent
```

```{r}
# get summary of each univariate model
TN_lm_uni_gap_summary <- lapply(TN_lm_uni_gap, summary)

# add to dataframe
TN_uni_gap_summary_all <- Reduce(full_join,lapply(mapply(cbind, lapply(lapply(TN_lm_uni_gap_summary, coef), as.data.frame), "model"=names(TN_lm_uni_gap_summary), SIMPLIFY=F), rownames_to_column, var = "predictor")) 
TN_uni_gap_summary_all <- TN_uni_gap_summary_all[,c(5,6,1:4)]
TN_uni_gap_summary_all
```

## reduced
```{r}
# find possible reduced model
TN_lm_full_gap2 <- glm(formula = gap ~ FHT + EVIamp + EVImed + ELE + SLO + ASP + LGS + Moose_new, data = c_env_sites_scale_TN_gap, family = binomial(link = "logit"), method = "brglmFit") 

TN_lm_red_gap <- step(TN_lm_full_gap2, direction = "backward", trace = F)
```

```{r}
# create additional models
TN_lm_red_gap2 <- glm(gap ~ FHT + EVIamp +  ASP , data = c_env_sites_scale_TN_gap,
                       family = binomial("logit"), method = "brglmFit")
TN_lm_red_gap3 <- glm(gap ~ FHT +EVIamp +  LGS , data = c_env_sites_scale_TN_gap,
                       family = binomial("logit"), method = "brglmFit")
TN_lm_red_gap4 <- glm(gap ~ FHT +EVIamp, data = c_env_sites_scale_TN_gap,
                       family = binomial("logit"), method = "brglmFit")
TN_lm_red_gap5 <- glm(gap ~ FHT + LGS, data = c_env_sites_scale_TN_gap,
                       family = binomial("logit"), method = "brglmFit") 
TN_lm_red_gap6 <- glm(gap ~ FHT + ASP, data = c_env_sites_scale_TN_gap,
                       family = binomial("logit"), method = "brglmFit") 

# add together to compare
TN_red_gap_models_r <- list()
TN_red_gap_models_r[[names(TN_lm_uni_gap)[1]]] <- TN_lm_uni_gap[[1]]
TN_red_gap_models_r[[names(TN_lm_uni_gap)[2]]] <- TN_lm_uni_gap[[2]]
TN_red_gap_models_r[[names(TN_lm_uni_gap)[6]]] <- TN_lm_uni_gap[[6]]
TN_red_gap_models_r[[names(TN_lm_uni_gap)[7]]] <- TN_lm_uni_gap[[7]]
TN_red_gap_models_r[["TN_lm_red_gap"]] <- TN_lm_red_gap
TN_red_gap_models_r[["TN_lm_red_gap2"]] <- TN_lm_red_gap2
TN_red_gap_models_r[["TN_lm_red_gap3"]] <- TN_lm_red_gap3
TN_red_gap_models_r[["TN_lm_red_gap4"]] <- TN_lm_red_gap4
TN_red_gap_models_r[["TN_lm_red_gap5"]] <- TN_lm_red_gap5
TN_red_gap_models_r[["TN_lm_red_gap6"]] <- TN_lm_red_gap6
Modnames_TN_red <- paste(names(TN_red_gap_models_r))

TN_N <- length(c_env_sites_scale_TN_gap$unique_id)

R2_TN_gap_red <- data.frame(lapply(TN_red_gap_models_r, NagelkerkeR2_own, null = TN_lm_null_gap, N= TN_N)) %>% 
  pivot_longer(cols = c(1:10), values_to = "R2", names_to = "Modnames")

## Log likelihood
LL_TN_gap_red <- data.frame(lapply(lapply(TN_red_gap_models_r, logLik), function(x) x[[1]])) %>% 
  pivot_longer(cols = c(1:10), values_to = "LL", names_to = "Modnames")

## AICc
AICc_TN_gap_red <- data.frame(lapply(TN_red_gap_models_r, AICc)) %>% 
  pivot_longer(cols = c(1:10), values_to = "AICc", names_to = "Modnames") 

## put together
aic_R2_LL_TN_red_gap <- full_join(LL_TN_gap_red, AICc_TN_gap_red) %>% 
  full_join(R2_TN_gap_red) 

```

## compare
```{r}
TN_models_gap_r <- list()

TN_models_gap_r[["TN_lm_null_gap"]] <- TN_lm_null_gap
TN_models_gap_r[["TN_lm_full_gap"]] <- TN_lm_full_gap
TN_models_gap_r[[names(TN_lm_uni_gap)[1]]] <- TN_lm_uni_gap[[1]]
TN_models_gap_r[[names(TN_lm_uni_gap)[2]]] <- TN_lm_uni_gap[[2]]
TN_models_gap_r[[names(TN_lm_uni_gap)[3]]] <- TN_lm_uni_gap[[3]]
TN_models_gap_r[[names(TN_lm_uni_gap)[4]]] <- TN_lm_uni_gap[[4]]
TN_models_gap_r[[names(TN_lm_uni_gap)[5]]] <- TN_lm_uni_gap[[5]]
TN_models_gap_r[[names(TN_lm_uni_gap)[6]]] <- TN_lm_uni_gap[[6]]
TN_models_gap_r[[names(TN_lm_uni_gap)[7]]] <- TN_lm_uni_gap[[7]]
TN_models_gap_r[[names(TN_lm_uni_gap)[8]]] <- TN_lm_uni_gap[[8]]
TN_models_gap_r[["TN_lm_red_gap"]] <- TN_lm_red_gap
TN_models_gap_r[["TN_lm_red_gap2"]] <- TN_lm_red_gap2
TN_models_gap_r[["TN_lm_red_gap3"]] <- TN_lm_red_gap3
TN_models_gap_r[["TN_lm_red_gap4"]] <- TN_lm_red_gap4
TN_models_gap_r[["TN_lm_red_gap5"]] <- TN_lm_red_gap5
TN_models_gap_r[["TN_lm_red_gap6"]] <- TN_lm_red_gap6

Modnames_TN_gap_r <- paste(names(TN_models_gap_r)) # make sure names are same as model name

LL_TN_gap <- data.frame(lapply(lapply(TN_models_gap_r, logLik), function(x) x[[1]])) %>% 
  pivot_longer(cols = c(1:16), values_to = "LL", names_to = "Modnames")

R2_TN_gap <- data.frame(lapply(TN_models_gap_r, NagelkerkeR2_own, null = TN_lm_null_gap, N= TN_N)) %>% 
  pivot_longer(cols = c(1:16), values_to = "R2", names_to = "Modnames")

AICc_TN_gap <- data.frame(lapply(TN_models_gap_r, AICc)) %>% 
  pivot_longer(cols = c(1:16), values_to = "AICc", names_to = "Modnames") 

aic_R2_LL_TN_gap <- full_join(LL_TN_gap, AICc_TN_gap) %>% 
  full_join(R2_TN_gap)
```
