==> 9) gap cross validation.Rmd <==
---
title: "9) gap cross validation"
author: "Emmerson Wilson"
date: "2024-06-17"
output: html_document
---

```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  cv,
  lme4,
  terra,
  boot,
  setRNG,
  brglm2,
  stringr)
```


```{r}
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/c_gap_predictions")
```

# import training datasets
```{r}
# made in 5) c lm total.Rmd
GM_gap_env <- read_xlsx("created data/c_env_sites_scale_GM_gap.xlsx")
TN_gap_env <- read_xlsx("created data/c_env_sites_scale_TN_gap.xlsx")
```

# GM

## top models
```{r}
# list of variables in best models from 5) c lm total.Rmd
# run them
var_GM_bi_red <- list("FHT", "EVIamp")
GM_gap_red <- firths_glm(var_GM_bi_red, df = GM_gap_env, c = "gap")

GM_var_lists_gap <- list(var_GM_bi_red)
GM_gap_top_models <- list(GM_gap_red)
```

## cross validation on top mdoels
```{r}
k <- 10
GM_cv_gap_Output <- NULL

for(x in seq(1,100,1)){
GM_gap_cv <- lapply(GM_gap_top_models, cv.glm, data=GM_gap_env, K = k)
names(GM_gap_cv) <- GM_var_lists_gap

GM_gap_cv_delta <- lapply(GM_gap_cv, `[`, c('delta'))

GM_gap_cv_delta <- do.call(rbind.data.frame, GM_gap_cv_delta)
GM_gap_cv_delta$names <- rownames(GM_gap_cv_delta)

GM_cv_gap_Output <- rbind(GM_cv_gap_Output, data.frame(Iteration = x, model = GM_gap_cv_delta$names, delta = GM_gap_cv_delta$delta))
}

GM_cv_gap_Output_delta1_2 <- GM_cv_gap_Output %>% 
  mutate(model_2 = substr(model,1,nchar(model)-2)) %>% 
  mutate(n = paste0("delta_",str_sub(model, -1, -1)))  %>% 
  select(!2) %>% 
  pivot_wider(values_from = delta, names_from = n)
```

## cross validation of randomized data
```{r}
GM_gap_cvrand <- lapply(GM_var_lists_gap, FUN = gap_cv.brgml, d = GM_gap_env, k = k)
names(GM_gap_cvrand) <- GM_var_lists_gap

GM_gap_cvrand_95 <- lapply(GM_gap_cvrand, `[`, c('conf.int'))

GM_gap_cvrand_95 <- do.call(rbind.data.frame, GM_gap_cvrand_95)
GM_gap_cvrand_95$model <- rownames(GM_gap_cvrand_95)

GM_gap_cvrand_95_1_2 <- GM_gap_cvrand_95%>% 
  mutate(model_2 = substr(model,1,nchar(model)-2)) %>% 
  mutate(n = paste0("ci95_",str_sub(model, -1, -1)))  %>% 
  select(!2) %>% 
  pivot_wider(values_from = conf.int, names_from = n)
```

## number of times models validate out of 100
```{r}
GM_val_cv_gap <- GM_cv_gap_Output_delta1_2 %>% 
  full_join(GM_gap_cvrand_95_1_2) %>% 
  mutate(diff = ci95_1 - delta_1) %>% 
  mutate(validate = ifelse(diff<0.0, "no", "yes"))
GM_nval_cv_gap <- GM_val_cv_gap %>% 
  group_by(model_2) %>% 
  summarize(n_y = sum(validate == "yes"), n_n = sum(validate == "no"))
```

# TN

## top models
```{r}
# list of variables in best models from 5) c lm total.Rmd
# run them
var_TN_bi_red <- list("FHT", "EVIamp", "ASP", "LGS")
TN_gap_red <- firths_glm(var_TN_bi_red, df = TN_gap_env, c = "gap")

var_TN_bi_red2 <- list("FHT", "EVIamp", "LGS")
TN_gap_red2 <- firths_glm(var_TN_bi_red2, df = TN_gap_env, c = "gap")

var_TN_bi_red3 <- list("FHT", "EVIamp", "ASP")
TN_gap_red3 <- firths_glm(var_TN_bi_red3, df = TN_gap_env, c = "gap")

var_TN_bi_red4 <- list("FHT", "EVIamp")
TN_gap_red4 <- firths_glm(var_TN_bi_red4, df = TN_gap_env, c = "gap")


TN_var_lists_gap <- list(var_TN_bi_red, var_TN_bi_red2, var_TN_bi_red3, var_TN_bi_red4)
TN_gap_top_models <- list(TN_gap_red, TN_gap_red2, TN_gap_red3, TN_gap_red4)
```

## cross validation on top mdoels
```{r}
k <- 9
TN_cv_gap_Output <- NULL

for(x in seq(1,100,1)){
TN_gap_cv <- lapply(TN_gap_top_models, cv.glm, data=TN_gap_env, K = k)
names(TN_gap_cv) <- TN_var_lists_gap

TN_gap_cv_delta <- lapply(TN_gap_cv, `[`, c('delta'))

TN_gap_cv_delta <- do.call(rbind.data.frame, TN_gap_cv_delta)
TN_gap_cv_delta$names <- rownames(TN_gap_cv_delta)

TN_cv_gap_Output <- rbind(TN_cv_gap_Output, data.frame(Iteration = x, model = TN_gap_cv_delta$names, delta = TN_gap_cv_delta$delta))
}

TN_cv_gap_Output_delta1_2 <- TN_cv_gap_Output %>% 
  mutate(model_2 = substr(model,1,nchar(model)-2)) %>% 
  mutate(n = paste0("delta_",str_sub(model, -1, -1)))  %>% 
  select(!2) %>% 
  pivot_wider(values_from = delta, names_from = n)
```

## cross validation of randomized data
```{r}
TN_gap_cvrand <- lapply(TN_var_lists_gap, FUN = gap_cv.brgml, d = TN_gap_env, k = k)
names(TN_gap_cvrand) <- TN_var_lists_gap

TN_gap_cvrand_95 <- lapply(TN_gap_cvrand, `[`, c('conf.int'))

TN_gap_cvrand_95 <- do.call(rbind.data.frame, TN_gap_cvrand_95)
TN_gap_cvrand_95$model <- rownames(TN_gap_cvrand_95)

TN_gap_cvrand_95_1_2 <- TN_gap_cvrand_95%>% 
  mutate(model_2 = substr(model,1,nchar(model)-2)) %>% 
  mutate(n = paste0("ci95_",str_sub(model, -1, -1)))  %>% 
  select(!2) %>% 
  pivot_wider(values_from = conf.int, names_from = n)
```

## number of times models validate out of 100
```{r}
TN_val_cv_gap <- TN_cv_gap_Output_delta1_2 %>% 
  full_join(TN_gap_cvrand_95_1_2) %>% 
  mutate(diff = ci95_1 - delta_1) %>% 
  mutate(validate = ifelse(diff<0.0, "no", "yes"))
TN_nval_cv_gap <- TN_val_cv_gap %>% 
  group_by(model_2) %>% 
  summarize(n_y = sum(validate == "yes"), n_n = sum(validate == "no"))
```
