==> 10) predictable space.Rmd <==
---
title: "7) predictable space"
author: "Emmerson Wilson"
date: "2024-06-17"
output: html_document
---

```{r}
pacman::p_load(
  terra,
  tidyverse,
  readxl)
```


```{r}
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/c_gap_predictions")
```

# import rasters
```{r}
# stacked rasters of enviornmental data
GM_env_rast <- rast("required data/GM_env_raster.tif") 
TN_env_rast <- rast("required data/TN_env_raster.tif") 

# layer to clip by: cut out all human use areas, and space where values of environmental variables in top models are outside of range sampled, all wetlands
clip_GM_rast <- rast("required data/GM_predict_binary.tif") 
clip_TN_rast <- rast("required data/TN_predict_binary.tif")
```

## prep clipping space
```{r}
# change anywhere equal to 0 into NA
clip_GM_rast[clip_GM_rast==0] <- NA

clip_TN_rast[clip_TN_rast==0] <- NA
```

## clip
```{r}
GM_env_rast_masked <- crop(GM_env_rast, clip_GM_rast, mask=TRUE) 
names(GM_env_rast_masked) <-  c( "SPC", "FAC", "FHT", "CC", "LCT", "EVIamp", "EVImed","SLO", "ELE", "ASP", "LGS", "Moose_new",  "Moose_newl",  "Moose_newh",  "Moose_newx") # name layers

#remove unwanted layers
GM_env_rast_masked <- subset(GM_env_rast_masked, c(3,6:12))
 #check

TN_env_rast_masked <- crop(TN_env_rast, clip_TN_rast, mask=TRUE)
names(TN_env_rast_masked) <-  c( "SPC", "FAC", "FHT", "CC", "LCT", "EVIamp", "EVImed","SLO", "ELE", "ASP", "LGS", "Moose_new",  "Moose_newl",  "Moose_newh",  "Moose_newx") # name layers

#remove unwanted layers
TN_env_rast_masked <- subset(TN_env_rast_masked, c(3,6:12))
```

# rescale variables
## rescaled data from plot locations
```{r}
env_for_scale_GM <- read_xlsx("required data/c_m2_site.xlsx") %>%
  full_join(read_xlsx("required data/env_sites_final.xlsx")) %>% 
  filter(park_id == "GM") %>% 
  dplyr::select(!c(FAC, SPC, CC, CEC_LCT)) %>% 
  dplyr::select(c(11:13, 16, 15, 14, 18)) # make sure in same order as raster stack

env_for_scale_TN <- read_xlsx("required data/c_m2_site.xlsx") %>%
  full_join(read_xlsx("required data/env_sites_final.xlsx")) %>% 
  filter(park_id == "TN") %>% 
  dplyr::select(!c(FAC, SPC, CC, CEC_LCT)) %>% 
  dplyr::select(c(11:13, 16, 15, 14, 18))

# scale and save the values each variable was scaled by
envscale_GM <- attr(scale(env_for_scale_GM, center = F), "scaled:scale") 
envscale_TN <- attr(scale(env_for_scale_TN, center = F), "scaled:scale") 
```

## apply to entire raster
```{r}
# make rasters with proper dimensions to write over
GM_env_raster_masked_scale <- GM_env_rast_masked
GM_env_raster_masked_nm <- subset(GM_env_rast_masked, c(1:7)) # remove moose from stack of rasters to be scaled

TN_env_raster_masked_scale <- TN_env_rast_masked
TN_env_raster_masked_nm <- subset(TN_env_rast_masked, c(1:7))

# rescale GM
for (i in 1:nlyr(GM_env_raster_masked_nm)){
  sub <- subset(GM_env_raster_masked_nm,i)
  GM_env_raster_masked_scale[[sub@cpp$names]] <- terra::scale(sub, center = FALSE, scale = envscale_GM[[i]]) # need to change cpp to ptr if working on windows
  }

# rescale TN
for (i in 1:nlyr(TN_env_raster_masked_nm)){
  sub <- subset(TN_env_raster_masked_nm,i)
  TN_env_raster_masked_scale[[sub@cpp$names]] <- terra::scale(sub, center = FALSE, scale = envscale_TN[[i]]) # need to change cpp to ptr if working on windows
  }
```

##export
```{r}
# save
writeRaster(GM_env_raster_masked_scale, filename="created data/GM_env_raster_masked_scaled.tif", overwrite=TRUE)
writeRaster(TN_env_raster_masked_scale, filename="created data/TN_env_raster_masked_scaled.tif", overwrite=TRUE)
```
