==> 11) c predictions.Rmd <==
---
title: "8) c predictions"
author: "Emmerson Wilson"
date: "2024-06-17"
output: html_document
---

```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  cv,
  lme4,
  terra)
```

```{r}
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/c_gap_predictions")
```

# import training datasets
```{r}
GM_c_env <- read_xlsx("created data/c_env_sites_scale_GM.xlsx")
TN_c_env <- read_xlsx("created data/c_env_sites_scale_TN.xlsx")
```

# import scaled environmental rasters
```{r}
GM_env_raster_masked_scale <- rast("created data/GM_env_raster_masked_scaled.tif")
TN_env_raster_masked_scale <- rast("created data/TN_env_raster_masked_scaled.tif")
```

# import glm function
```{r}
source("4) glm functions.R")
```

# predict

## GM
```{r}
# list of variables in best model from 5) c lm total.Rmd
GM_var <- list("FHT", "EVIamp")

# run best model
GM_model_tot <- lm_glm(GM_var, df = GM_c_env, c = "c_m2_site")
summary(GM_model_tot)

# predict
GM_predict_tot <- terra::predict(GM_env_raster_masked_scale, GM_model_tot)

# transform predictions
GM_predict_tot_transform <- GM_predict_tot
values(GM_predict_tot_transform) <- exp(values(GM_predict_tot))

# export
writeRaster(GM_predict_tot_transform, filename="created data/GM_predict_tot.tif", overwrite=TRUE)
```

## TN
```{r}
# list of variables in best model from 5) c lm total.Rmd
TN_var <- list("FHT", "EVIamp")

# run best model
TN_model_tot <- lm_glm(TN_var, df = subset(TN_c_env), c = "c_m2_site")
summary(TN_model_tot)

# predict
TN_predict_tot <- terra::predict(TN_env_raster_masked_scale, TN_model_tot)

# transform predictions
TN_predict_tot_transform <- TN_predict_tot
values(TN_predict_tot_transform) <- exp(values(TN_predict_tot))

# export
writeRaster(TN_predict_tot_transform, filename="created data/TN_predict_tot.tif", overwrite=TRUE)
```
